+++ pepys_import/import.py
@@ -3,7 +3,7 @@ import argparse
 
 from pepys_import.file.file_processor import FileProcessor
 from pepys_import.core.store.data_store import DataStore
-from pepys_import.file.get_importers import get_importers
+from importers.get_importers import get_importers
 
 FILE_PATH = os.path.abspath(__file__)
 DIRECTORY_PATH = os.path.dirname(FILE_PATH)
