+++ tests/test_load_rep.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from pepys_import.file.replay_importer import ReplayImporter
+from importers.replay_importer import ReplayImporter
 from pepys_import.file.file_processor import FileProcessor
 from pepys_import.core.store.data_store import DataStore
